Move datamodels.open() from __init__.py to util.py, so open() can be used with a relative import within the package.  API remains the same.
