+++ jwst/datamodels/util.py
@@ -4,10 +4,135 @@ Various utility functions and data types
 from __future__ import absolute_import, unicode_literals, division, print_function
 
 import sys
-
+from os.path import basename
 import numpy as np
-
 from astropy.extern import six
+from astropy.io import fits
+
+from . import model_base
+
+
+def open(init=None, extensions=None):
+    """
+    Creates a DataModel from a number of different types
+
+    Parameters
+    ----------
+
+    init : shape tuple, file path, file object, astropy.io.fits.HDUList, 
+           numpy array, dict, None
+
+        - None: A default data model with no shape
+
+        - shape tuple: Initialize with empty data of the given shape
+
+        - file path: Initialize from the given file (FITS , JSON or ASDF)
+
+        - readable file object: Initialize from the given file object
+
+        - astropy.io.fits.HDUList: Initialize from the given
+          `~astropy.io.fits.HDUList`
+
+        - A numpy array: A new model with the data array initialized
+          to what was passed in.
+
+        - dict: The object model tree for the data model
+
+    extensions : list of AsdfExtension
+        A list of extensions to the ASDF to support when reading
+        and writing ASDF files.
+
+   Results
+    -------
+
+    model : DataModel instance
+    """
+
+    if init is None:
+        return model_base.DataModel(None)
+    # Send _asn.json files to ModelContainer; avoid shape "cleverness" below
+    elif (isinstance(init, six.string_types) and
+            basename(init).split('.')[0].split('_')[-1] == 'asn'):
+        try:
+            from . import container
+            return container.ModelContainer(init, extensions=extensions)
+        except:
+            raise TypeError(
+                "init ASN not valid for ModelContainer"
+                )
+    elif isinstance(init, model_base.DataModel):
+        # Copy the object so it knows not to close here
+        return init.__class__(init)
+    elif isinstance(init, tuple):
+        for item in init:
+            if not isinstance(item, int):
+                raise ValueError("shape must be a tuple of ints")
+        shape = init
+    elif isinstance(init, np.ndarray):
+        shape = init.shape
+    else:
+        if isinstance(init, (six.text_type, bytes)) or hasattr(init, "read"):
+            hdulist = fits.open(init)
+        elif isinstance(init, fits.HDUList):
+            hdulist = init
+        else:
+            raise TypeError(
+                "init must be None, shape tuple, file path, "
+                "readable file object, or astropy.io.fits.HDUList")
+
+        shape = ()
+        try:
+            hdu = hdulist[(fits_header_name('SCI'), 1)]
+        except KeyError:
+            pass
+        else:
+            if hasattr(hdu, 'shape'):
+                shape = hdu.shape
+
+    # Be clever about which type to return, otherwise, just return a new 
+    # instance of the requested class
+    if len(shape) == 0:
+        new_class = model_base.DataModel
+    elif len(shape) == 4:
+        # It's a RampModel, MIRIRampModel, or QuadModel
+        try:
+            dqhdu = hdulist[fits_header_name('DQ')]
+        except KeyError:
+            # It's a RampModel or MIRIRampModel
+            try:
+                refouthdu = hdulist[fits_header_name('REFOUT')]
+            except KeyError:
+                # It's a RampModel
+                from . import ramp
+                new_class = ramp.RampModel
+            else:
+                # It's a MIRIRampModel
+                from . import miri_ramp
+                new_class = miri_ramp.MIRIRampModel
+        else:
+            # It's a QuadModel
+            from . import quad
+            new_class = quad.QuadModel
+    elif len(shape) == 3:
+        # It's a CubeModel
+        from . import cube
+        new_class = cube.CubeModel
+    elif len(shape) == 2:
+        try:
+            hdu = hdulist[(fits_header_name('SCI'), 2)]
+        except (KeyError, NameError):
+            # It's an ImageModel
+            from . import image
+            new_class = image.ImageModel
+        else:
+            # It's a MultiSlitModel
+            from . import multislit
+            new_class = multislit.MultiSlitModel
+    else:
+        raise ValueError("Don't have a DataModel class to match the shape")
+
+    return new_class(init, extensions=extensions)
+
 
 def can_broadcast(a, b):
     """
